Adicionando JaCoCo e resolvendo pendencias de compilação
