+++ enock-airlines/enock-airlines.iml
@@ -1,37 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <module type="JAVA_MODULE" version="4">
-  <component name="NewModuleRootManager" inherit-compiler-output="true">
-    <exclude-output />
-    <content url="file://$MODULE_DIR$">
+  <component name="AdditionalModuleElements">
+    <content url="file://$MODULE_DIR$" dumb="true">
       <sourceFolder url="file://$MODULE_DIR$/src" isTestSource="false" />
     </content>
-    <orderEntry type="inheritedJdk" />
-    <orderEntry type="sourceFolder" forTests="false" />
-    <orderEntry type="module-library">
-      <library name="JUnit4">
-        <CLASSES>
-          <root url="jar://$MODULE_DIR$/../lib/junit-4.13.1.jar!/" />
-          <root url="jar://$MODULE_DIR$/../lib/hamcrest-core-1.3.jar!/" />
-        </CLASSES>
-        <JAVADOC />
-        <SOURCES />
-      </library>
-    </orderEntry>
-    <orderEntry type="module-library">
-      <library name="JUnit5.8.1">
-        <CLASSES>
-          <root url="jar://$MODULE_DIR$/../lib/junit-jupiter-5.8.1.jar!/" />
-          <root url="jar://$MODULE_DIR$/../lib/junit-jupiter-api-5.8.1.jar!/" />
-          <root url="jar://$MODULE_DIR$/../lib/opentest4j-1.2.0.jar!/" />
-          <root url="jar://$MODULE_DIR$/../lib/junit-platform-commons-1.8.1.jar!/" />
-          <root url="jar://$MODULE_DIR$/../lib/apiguardian-api-1.1.2.jar!/" />
-          <root url="jar://$MODULE_DIR$/../lib/junit-jupiter-params-5.8.1.jar!/" />
-          <root url="jar://$MODULE_DIR$/../lib/junit-jupiter-engine-5.8.1.jar!/" />
-          <root url="jar://$MODULE_DIR$/../lib/junit-platform-engine-1.8.1.jar!/" />
-        </CLASSES>
-        <JAVADOC />
-        <SOURCES />
-      </library>
-    </orderEntry>
   </component>
 </module>
