+++ enock-airlines/src/main/Reserva.java
@@ -9,11 +9,12 @@ public class Reserva {
         this.vooRepository = new VooRepository();
     }
 
-    public void reservarVoo(int idVoo, int qtdVagas, Passageiro passageiro) {
+    public String reservarVoo(int idVoo, int qtdVagas, Passageiro passageiro) {
         List<Voo> voos = this.vooRepository.getVoos();
         Voo voo = voos.get(idVoo);
 
         voo.reservaVagas(passageiro, qtdVagas);
+        return null;
     }
 
     public VooRepository getVooRepository() {
