+++ enock-airlines/src/test/tests/TDD/PassageiroTeste.java
@@ -1,9 +1,9 @@
-package tests.TDD;
+package TDD;
 import main.Passageiro;
 import org.junit.jupiter.api.Test;
 
-import static org.junit.Assert.assertTrue;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 public class PassageiroTeste {
     private Passageiro passageiro;
