+++ enock-airlines/src/test/tests/functionalTests/avl/ReservaTeste.java
@@ -1,4 +1,4 @@
-package tests.functionalTests.avl;
+package functionalTests.avl;
 
 import main.Passageiro;
 import main.Reserva;
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class ReservaTeste {
